# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Redocly FastAPI Sample
  description: Sample API for OpenAPI and Redocly experimentation. Includes CRUD-style
    endpoints for users, products, and orders.
  version: 0.1.0
paths:
  /users/:
    get:
      tags:
      - Users
      summary: List users
      description: Return all users.
      operationId: list_users_users__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/User'
                type: array
                title: Response List Users Users  Get
    post:
      tags:
      - Users
      summary: Create user
      description: Create a new user. Returns `400` if the email already exists.
      operationId: create_user_users__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Email already exists
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /users/{user_id}:
    get:
      tags:
      - Users
      summary: Get user
      description: Get a user by ID.
      operationId: get_user_users__user_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Users
      summary: Update user
      description: Replace a user by ID.
      operationId: update_user_users__user_id__put
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Users
      summary: Delete user
      description: Delete a user by ID.
      operationId: delete_user_users__user_id__delete
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      responses:
        '204':
          description: Successful Response
        '404':
          description: User not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /products/:
    get:
      tags:
      - Products
      summary: List products
      description: Return all products.
      operationId: list_products_products__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Product'
                type: array
                title: Response List Products Products  Get
    post:
      tags:
      - Products
      summary: Create product
      description: Create a new product.
      operationId: create_product_products__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
        required: true
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /products/{product_id}:
    get:
      tags:
      - Products
      summary: Get product
      description: Get a product by ID.
      operationId: get_product_products__product_id__get
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: integer
          title: Product Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Products
      summary: Update product
      description: Replace a product by ID.
      operationId: update_product_products__product_id__put
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: integer
          title: Product Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Products
      summary: Delete product
      description: Delete a product by ID.
      operationId: delete_product_products__product_id__delete
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: integer
          title: Product Id
      responses:
        '204':
          description: Successful Response
        '404':
          description: Product not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /orders/:
    get:
      tags:
      - Orders
      summary: List orders
      description: Return all orders.
      operationId: list_orders_orders__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Order'
                type: array
                title: Response List Orders Orders  Get
    post:
      tags:
      - Orders
      summary: Create order
      description: Create a new order for an existing user and products.
      operationId: create_order_orders__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
        required: true
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: User or product not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /orders/{order_id}:
    get:
      tags:
      - Orders
      summary: Get order
      description: Get an order by ID.
      operationId: get_order_orders__order_id__get
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: integer
          title: Order Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Orders
      summary: Update order
      description: Replace an order by ID.
      operationId: update_order_orders__order_id__put
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: integer
          title: Order Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order, user, or product not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Orders
      summary: Delete order
      description: Delete an order by ID.
      operationId: delete_order_orders__order_id__delete
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: integer
          title: Order Id
      responses:
        '204':
          description: Successful Response
        '404':
          description: Order not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Order:
      properties:
        user_id:
          type: integer
          title: User Id
          description: User placing the order
          example: 1
        items:
          items:
            $ref: '#/components/schemas/OrderItem'
          type: array
          minItems: 1
          title: Items
        id:
          type: integer
          title: Id
          description: Unique order ID
          example: 1
        total:
          type: number
          title: Total
          description: Computed order total in USD
          example: 389.97
      type: object
      required:
      - user_id
      - items
      - id
      - total
      title: Order
      example:
        items:
        - product_id: 1
          quantity: 2
        - product_id: 2
          quantity: 1
        user_id: 1
    OrderCreate:
      properties:
        user_id:
          type: integer
          title: User Id
          description: User placing the order
          example: 1
        items:
          items:
            $ref: '#/components/schemas/OrderItem'
          type: array
          minItems: 1
          title: Items
      type: object
      required:
      - user_id
      - items
      title: OrderCreate
      example:
        items:
        - product_id: 1
          quantity: 2
        - product_id: 2
          quantity: 1
        user_id: 1
    OrderItem:
      properties:
        product_id:
          type: integer
          title: Product Id
          description: Referenced product ID
          example: 1
        quantity:
          type: integer
          exclusiveMinimum: 0.0
          title: Quantity
          description: Number of items
          example: 2
      type: object
      required:
      - product_id
      - quantity
      title: OrderItem
    Product:
      properties:
        name:
          type: string
          title: Name
          description: Product name
          example: Mechanical Keyboard
        description:
          type: string
          title: Description
          example: A tactile mechanical keyboard
        price:
          type: number
          exclusiveMinimum: 0.0
          title: Price
          description: Price in USD
          example: 129.99
        id:
          type: integer
          title: Id
          description: Unique product ID
          example: 1
      type: object
      required:
      - name
      - description
      - price
      - id
      title: Product
      example:
        description: A tactile mechanical keyboard
        name: Mechanical Keyboard
        price: 129.99
    ProductCreate:
      properties:
        name:
          type: string
          title: Name
          description: Product name
          example: Mechanical Keyboard
        description:
          type: string
          title: Description
          example: A tactile mechanical keyboard
        price:
          type: number
          exclusiveMinimum: 0.0
          title: Price
          description: Price in USD
          example: 129.99
      type: object
      required:
      - name
      - description
      - price
      title: ProductCreate
      example:
        description: A tactile mechanical keyboard
        name: Mechanical Keyboard
        price: 129.99
    User:
      properties:
        name:
          type: string
          title: Name
          description: Full name of the user
          example: Alice Johnson
        email:
          type: string
          format: email
          title: Email
          description: Unique email
          example: alice@example.com
        id:
          type: integer
          title: Id
          description: Unique user ID
          example: 1
      type: object
      required:
      - name
      - email
      - id
      title: User
      example:
        email: alice@example.com
        name: Alice Johnson
    UserCreate:
      properties:
        name:
          type: string
          title: Name
          description: Full name of the user
          example: Alice Johnson
        email:
          type: string
          format: email
          title: Email
          description: Unique email
          example: alice@example.com
      type: object
      required:
      - name
      - email
      title: UserCreate
      example:
        email: alice@example.com
        name: Alice Johnson
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
tags:
- name: Users
  description: Operations related to user records and profiles.
- name: Products
  description: Operations related to product catalog management.
- name: Orders
  description: Operations related to order lifecycle and totals.